+++ src/Form/LegacyConsumer/Commands/SetupFieldEmailTag.php
@@ -2,6 +2,9 @@
 
 namespace Give\Form\LegacyConsumer\Commands;
 
+use Give\Framework\FieldsAPI\Concerns\HasEmailTag;
+use Give\Framework\FieldsAPI\Concerns\HasLabel;
+use Give\Framework\FieldsAPI\Concerns\StoreAsMeta;
 use Give\Framework\FieldsAPI\Field;
 use Give\Framework\FieldsAPI\Group;
 
@@ -33,6 +36,7 @@ class SetupFieldEmailTag {
 	 * @return void
 	 */
 	public function register( Field $field ) {
+		/** @var Field|HasLabel|HasEmailTag|StoreAsMeta $field */
 		give_add_email_tag(
 			[
 				'tag'      => $field->getEmailTag() ?: $field->getName(), // The tag name.