+++ src/Form/LegacyConsumer/templates/textarea.html.php
@@ -1,7 +1,9 @@
 <?php /** @var Give\Framework\FieldsAPI\Textarea $field */ ?>
+<?php /** @var string $fieldIdAttribute */ ?>
 <textarea
-	name="give_<?php echo $field->getName(); ?>"
-	id="give-<?php echo $field->getName(); ?>"
+	name="<?php echo $field->getName(); ?>"
+	id="<?php echo $fieldIdAttribute; ?>"
 	<?php echo $field->isRequired() ? 'required' : ''; ?>
 	<?php echo $field->isReadOnly() ? 'readonly' : ''; ?>
+	<?php echo ( $maxLength = $field->getMaxLength() ) ? "maxlength=\"$maxLength\"" : ''; ?>
 ></textarea>