+++ src/Framework/FieldsAPI/Email.php
@@ -4,6 +4,7 @@ namespace Give\Framework\FieldsAPI;
 
 /**
  * @since 2.12.0
+ * @since 2.14.0 add min/max length validation
  */
 class Email extends Field {
 
@@ -11,6 +12,8 @@ class Email extends Field {
 	use Concerns\HasEmailTag;
 	use Concerns\HasHelpText;
 	use Concerns\HasLabel;
+	use Concerns\HasMaxLength;
+	use Concerns\HasMinLength;
 	use Concerns\HasPlaceholder;
 	use Concerns\ShowInReceipt;
 	use Concerns\StoreAsMeta;