+++ src/Framework/FieldsAPI/Textarea.php
@@ -4,6 +4,7 @@ namespace Give\Framework\FieldsAPI;
 
 /**
  * @since 2.12.0
+ * @since 2.14.0 Add support for min/max length
  */
 class Textarea extends Field {
 
@@ -13,6 +14,8 @@ class Textarea extends Field {
 	use Concerns\HasPlaceholder;
 	use Concerns\ShowInReceipt;
 	use Concerns\StoreAsMeta;
+	use Concerns\HasMaxLength;
+	use Concerns\HasMinLength;
 
 	const TYPE = 'textarea';
 }