+++ src/Views/Form/Templates/Sequoia/assets/js/form.js
@@ -128,17 +128,6 @@
 				$( '#give_checkout_user_info' ).after( $( '.give-fee-recovery-donors-choice' ) );
 			}
 			navigator.goToStep( getInitialStep() );
-
-			// Fields API: Run setup for custom checkbox fields.
-			const customCheckboxes = document.querySelectorAll( '[data-field-type="checkbox"]' );
-			Array.from( customCheckboxes ).forEach( ( el ) => {
-				const containerSelector = '[data-field-name="' + el.getAttribute( 'data-field-name' ) + '"]';
-				setupCheckbox( {
-					container: containerSelector,
-					label: containerSelector + ' label',
-					input: containerSelector + ' input[type="checkbox"]',
-				} );
-			} );
 		},
 		back: () => {
 			const prevStep = navigator.currentStep !== 0 ? navigator.currentStep - 1 : 0;
@@ -456,19 +445,16 @@
 	$( document ).on( 'give_gateway_loaded', refreshPersonalInformationSection );
 
 	// Setup fields.
+	setupLegacyConsumerCheckboxAndRadio();
 	setupSelectInputs();
 	setupRegistrationFormInputFields();
 	setupFFMInputs();
 	setupInputIcons();
 
-	if ( 'enabled' === templateOptions.payment_amount.decimals_enabled ) {
-		updateFormDonationLevelsLabels();
-	}
-
 	/**
 	 * Limited scope of optional input labels, specifically to User Info, see issue #5160.
 	 */
-	setupOptionalInputLables(
+	setupOptionalInputLabels(
 		Array.from( document.querySelectorAll( '#give_checkout_user_info input[type="text"]' ) )
 	);
 
@@ -479,7 +465,7 @@
 	 *
 	 * @param {array} inputs An iteratable list of input elements.
 	 */
-	function setupOptionalInputLables( inputs ) {
+	function setupOptionalInputLabels( inputs ) {
 		inputs.filter( function( input ) {
 			return ! input.required;
 		} ).map( function( input ) {
@@ -574,6 +560,28 @@
 				}
 			}
 		} );
+	}
+
+	/**
+	 * Handle updating label classes for FFM radios and checkboxes
+	 *
+	 * @since 2.7.0
+	 * @param {object} evt Reference to FFM input element click event
+	 */
+	function handleFFMInput( evt ) {
+		if ( $( evt.target ).is( 'input' ) ) {
+			switch ( $( evt.target ).prop( 'type' ) ) {
+				case 'checkbox': {
+					$( evt.target ).closest( 'label' ).toggleClass( 'checked' );
+					break;
+				}
+				case 'radio': {
+					$( evt.target ).closest( 'label' ).addClass( 'selected' );
+					$( evt.target ).parent().siblings().removeClass( 'selected' );
+					break;
+				}
+			}
+		}
 	}
 
 	/**
@@ -759,6 +767,33 @@
 	}
 
 	/**
+	 * Setup prominent checkboxes (field api) (that use persistent borders on select)
+	 *
+	 * @since 2.14.0
+	 */
+	function setupLegacyConsumerCheckboxAndRadio(){
+		const customCheckboxes = document.querySelectorAll( '[data-field-type="checkbox"]' );
+		const customRadios = document.querySelectorAll( '[data-field-type="radio"] input' );
+		Array.from( customCheckboxes ).forEach( ( el ) => {
+			const containerSelector = '[data-field-name="' + el.getAttribute( 'data-field-name' ) + '"]';
+			setupCheckbox( {
+				container: containerSelector + ' label',
+				label: containerSelector + ' label',
+				input: containerSelector + ' input[type="checkbox"]',
+			} );
+		} );
+
+		Array.from( customRadios ).forEach( ( el ) => {
+			const uniqueInputSelector = `#${el.getAttribute( 'id' )}`;
+			const uniqueLabelSelector = `label[for=${el.getAttribute( 'id' )}]`;
+			setupRadio( {
+				label: uniqueLabelSelector,
+				input: uniqueInputSelector,
+			} );
+		} );
+	}
+
+	/**
 	 * Setup prominent checkboxes (that use persistent borders on select)
 	 *
 	 * @since 2.7.0
@@ -776,10 +811,32 @@
 				evt.stopPropagation();
 				evt.preventDefault();
 
-				$( input ).prop( 'checked', ! $( input ).prop( 'checked' ) );
+				$( input ).prop( 'checked', ! $( input ).prop( 'checked' ) ).focus();
 			}
 
 			$( container ).toggleClass( 'active' );
+		} );
+	}
+
+	/**
+	 * Handle updating label classes for FFM radios and checkboxes
+	 *
+	 * @since 2.7.0
+	 * @param {object} evt Reference to FFM input element click event
+	 */
+	function setupRadio( { label, input } ) {
+		// If checkbox is opted in by default, add border on load
+		if ( $( input ).prop( 'checked' ) === true ) {
+			$( label ).addClass( 'active' );
+		}
+
+		// Persist checkbox input border when selected
+		$( document ).on( 'click', label, function( evt ) {
+			evt.stopPropagation();
+
+			$( evt.target.parentElement ).find('label')
+				.not( evt.target ).removeClass( 'active' );
+			$( evt.target ).toggleClass( 'active' );
 		} );
 	}
 
@@ -805,28 +862,6 @@
 	 */
 	function getInitialStep() {
 		return Give.fn.getParameterByName( 'showDonationProcessingError' ) || Give.fn.getParameterByName( 'showFailedDonationError' ) ? 2 : 0;
-	}
-
-	/**
-	 * Handle updating label classes for FFM radios and checkboxes
-	 *
-	 * @since 2.7.0
-	 * @param {object} evt Reference to FFM input element click event
-	 */
-	function handleFFMInput( evt ) {
-		if ( $( evt.target ).is( 'input' ) ) {
-			switch ( $( evt.target ).prop( 'type' ) ) {
-				case 'checkbox': {
-					$( evt.target ).closest( 'label' ).toggleClass( 'checked' );
-					break;
-				}
-				case 'radio': {
-					$( evt.target ).closest( 'label' ).addClass( 'selected' );
-					$( evt.target ).parent().siblings().removeClass( 'selected' );
-					break;
-				}
-			}
-		}
 	}
 
 	function clearLoginNotices() {
@@ -868,36 +903,5 @@
 		if ( 'parentIFrame' in window ) {
 			window.parentIFrame.sendMessage( { action: 'giveScrollIframeInToView' } );
 		}
-	}
-
-	/**
-	 * Update decimal donation levels amount
-	 *
-	 * @since 2.11.0
-	 */
-	function updateFormDonationLevelsLabels() {
-		$( '.give-form' ).each( ( i, form ) => {
-			const donationForm = $( form );
-			const donationLevels = Give.form.fn.getVariablePrices( donationForm );
-			const symbol = Give.form.fn.getInfo( 'currency_symbol', donationForm );
-			const position = Give.form.fn.getInfo( 'currency_position', donationForm );
-			const precision = Give.form.fn.getInfo( 'number_decimals', donationForm );
-
-			$.each( donationLevels, function( j, level ) {
-				if ( 'custom' === level.price_id ) {
-					return;
-				}
-
-				const amount = Give.fn.numberHasDecimal( level.amount )
-					? Give.fn.formatCurrency( level.amount, { symbol, position, precision }, donationForm )
-					: level.amount;
-
-				const donationLevelLabel = '<div class="currency currency--' + position + '">' + symbol + '</div>' + amount;
-
-				donationForm
-					.find( '.give-btn-level-' + level.price_id  )
-					.html( donationLevelLabel );
-			} );
-		} );
 	}
 }( jQuery ) );